+++ formscriber/main.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -90,112 +88,6 @@ func index(res http.ResponseWriter, req *http.Request) {
  * the following are the REST request handlers
  */
 
-/* 2.1 GET Hep Articles
- * URL: http://localhost:8080/getArticles
- */
-type Articles struct {
-	Type string    `json:"type"`
-	D    []Article `json:"d"`
-}
-type Article struct {
-	ID       string   `json:"id,omitempty"`
-	Name     string   `json:"name,omitempty"`
-	Desc     string   `json:"description,omitempty"`
-	URL      string   `json:"URL"`
-	KeyWords *KeyWord `json:"keywords,omitempty"`
-}
-type KeyWord struct {
-	Name string `json:"name,omitempty"`
-}
-
-func GetArticleEndPoint(w http.ResponseWriter, request *http.Request) {
-	//	REST endpoint to get articles
-	jsonFile, err := os.Open("DATA/GetHelpArticleList.JSON")
-
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	// defer the closing of our jsonFile so that we can parse it later on
-	defer jsonFile.Close()
-
-	// read our opened JSON as a byte array.
-	byteValue, _ := ioutil.ReadAll(jsonFile)
-
-	// we initialize our Users array
-	var articles Articles
-	json.Unmarshal(byteValue, &articles)
-
-	json.NewEncoder(w).Encode(articles)
-}
-
-/* 2.2 GET Team
- * URL: http://localhost:8080/getTeam
- */
-type Team struct {
-	Type string       `json:"type"`
-	D    []TeamMember `json:"d"`
-}
-type TeamMember struct {
-	FName string `json:"fname"`
-	LName string `json:"lname"`
-	Role  string `json:"role"`
-}
-
-func GetTeamEndPoint(w http.ResponseWriter, request *http.Request) {
-	//	REST endpoint to get articles
-	jsonFile, err := os.Open("DATA/GetTeam.JSON")
-
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	// defer the closing of our jsonFile so that we can parse it later on
-	defer jsonFile.Close()
-
-	// read our opened JSON as a byte array.
-	byteValue, _ := ioutil.ReadAll(jsonFile)
-
-	// we initialize our Users array
-	var team Team
-	json.Unmarshal(byteValue, &team)
-
-	json.NewEncoder(w).Encode(team)
-}
-
-/* 2.3 GET HAQs
- * URL: http://localhost:8080/getFAQs
- */
-type FAQs struct {
-	Type string `json:"type"`
-	D    []FAQ  `json:"d"`
-}
-type FAQ struct {
-	Q string `json:"q"`
-	A string `json:"a"`
-}
-
-func GetFAQsEndPoint(w http.ResponseWriter, request *http.Request) {
-	//	REST endpoint to get articles
-	jsonFile, err := os.Open("DATA/GetFAQs.JSON")
-
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	// defer the closing of our jsonFile so that we can parse it later on
-	defer jsonFile.Close()
-
-	// read our opened JSON as a byte array.
-	byteValue, _ := ioutil.ReadAll(jsonFile)
-
-	// we initialize our Users array
-	var faqs FAQs
-	json.Unmarshal(byteValue, &faqs)
-
-	json.NewEncoder(w).Encode(faqs)
-}
-
 /* Main Code
  * THis is where everything is activated
  */
@@ -225,7 +117,7 @@ func main() {
 	http.Handle("/site/JS/", http.StripPrefix("/site/JS/", http.FileServer(http.Dir("site/JS/"))))
 	http.Handle("/site/IMG/", http.StripPrefix("/site/IMG/", http.FileServer(http.Dir("site/IMG"))))
 
-	http.HandleFunc("/getArticles", GetArticleEndPoint)
+	http.HandleFunc("/getArticles", GetArticlesEndPoint)
 	http.HandleFunc("/getTeam", GetTeamEndPoint)
 	http.HandleFunc("/getFAQs", GetFAQsEndPoint)
